Massive api rewrite
So completely changed the api of the library. Now user's create an Fzf
struct, call the run method, call the add_items method to add items to
the selection, then call the output method to get the output of fzf.

This makes a way more flexible api, allowing things like what i've done
in `workflows`, just 20 times easier.

Also added the convience fuction run_with_output which takes a vec of
items and a Fzf struct, and returns the result, which is the previous
api. Had to make this a seperate function as the output function is self
consuming on the Fzf struct